+++ src/core/dependencies.py
